+++ apps/server/src/core/events/activity/schemas/activity.schema.ts
@@ -1,8 +1,8 @@
 import { Prop, Schema, SchemaFactory } from "@nestjs/mongoose"
 import { Document, Schema as MongooseSchema } from "mongoose"
 
-@Schema({ versionKey: false, collection: "insights", timestamps: { createdAt: true, updatedAt: false } })
-export class Insights extends Document {
+@Schema({ versionKey: false, collection: "activities", timestamps: { createdAt: true, updatedAt: false } })
+export class Activity extends Document {
   @Prop({ type: MongooseSchema.Types.ObjectId, ref: "user", required: true })
   userId: MongooseSchema.Types.ObjectId
 
@@ -13,4 +13,4 @@ export class Insights extends Document {
   apiUri: string
 }
 
-export const InsightsSchema = SchemaFactory.createForClass(Insights)
+export const ActivitySchema = SchemaFactory.createForClass(Activity)