+++ apps/server/src/products/insights/queries/impl/get-events.query.ts
@@ -1,0 +1,5 @@
+export class GetEventsQuery {
+  constructor(
+    public readonly orgId: string,
+  ) { }
+}